requirements.txt
streamlit
pandas
numpy
altair
htbuilder
google-cloud-bigquery